refactor(iocuak-metadata): move modules out of domain folders
